usrspace/requirements.txt
cmd2==2.7.0
fire==0.7.1
markdown-it-py==4.0.0
mdurl==0.1.2
pendulum==3.1.0
Pygments==2.19.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.1.0
rich-argparse==1.7.1
six==1.17.0
termcolor==3.1.0
tzdata==2025.2
wcwidth==0.2.14
